integrated/webapp/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
async-timeout==4.0.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
decorator==4.4.2
Deprecated==1.2.13
Flask==2.2.2
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.22.2
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
moviepy==1.0.3
networkx==2.8.7
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyWavelets==1.4.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
SpeechRecognition==3.8.1
spotipy==2.20.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
v==0.0.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
